studentaanpassing

+++ src/main/java/be/vdab/oak3evaluationform/Service/StudentServiceImpl.java
@@ -25,7 +25,9 @@ public class StudentServiceImpl implements StudentService {
 
     @Override
     public int deleteStudent(Student student) {
-        studentJPARepository.delete(student);
+        if(studentJPARepository.exists(student.getId())){
+            studentJPARepository.delete(student);
+        studentJPARepository.delete(student);}
         return studentJPARepository.exists(student.getId())?1:0 ;
     }
